# Fix blobovnicza, metabase index improvements