1. Functions in `Update` and `Batch` should be idempotent.
2. Reduce allocations count during index processing in `Put` / `Delete`.