@alexvanin 2022-03-22
I see benchmark in this PR. Please, provide results of the bench w/ and w/o memory alloc optimization in comment below.