@fyrchik 2022-03-22
They are included in the commit messages. Here is a comparison of the master branch with a top commit:
```
dzeta@wpc ~/r/n/p/l/metabase (fix-blobstor)> benchstat before after
name              old time/op    new time/op    delta
Put/parallel-8      1.52ms ±14%    1.47ms ±14%   -3.16%  (p=0.030 n=20+19)
Put/sequential-8    5.29ms ± 2%    5.19ms ± 4%   -2.02%  (p=0.003 n=20+20)

name              old alloc/op   new alloc/op   delta
Put/parallel-8       134kB ±12%     118kB ± 4%  -11.47%  (p=0.000 n=20+17)
Put/sequential-8     173kB ± 1%     168kB ± 1%   -2.63%  (p=0.000 n=20+20)

name              old allocs/op  new allocs/op  delta
Put/parallel-8         573 ± 5%       468 ± 1%  -18.25%  (p=0.000 n=19+20)
Put/sequential-8       820 ± 1%       787 ± 1%   -3.97%  (p=0.000 n=20+19)
```